LayoutTests/fast/overflow/scroll-anchor-in-position-fixed-expected.txt

Tests scrolling to an anchor inside position:fixed doesn't try to scroll the page

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.scrollingElement.scrollTop is 800
PASS document.scrollingElement.scrollLeft is 100
PASS successfullyParsed is true

TEST COMPLETE
Anchor is here
